llm-celery/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aioredis==2.0.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.7.0
boto3==1.37.6
botocore==1.37.6
build==1.2.1
cachetools==5.3.3
celery==5.4.0
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
configobj==5.0.8
configparser==7.0.0
cryptography==42.0.5
dataclasses-json==0.6.4
datasets==3.3.2
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
etelemetry==0.3.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.1
fastapi-cli==0.0.4
filelock==3.13.4
filetype==1.2.0
fitz==0.0.1.dev2
Flask==3.0.3
flatbuffers==24.3.25
flower==2.0.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
gpt4all==2.8.2
greenlet==3.0.3
grpcio==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.29.2
humanfriendly==10.0
humanize==4.9.0
hyperframe==6.0.1
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
ipython==8.18.1
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.7
kubernetes==29.0.0
langchain==0.3.20
langchain-aws==0.2.15
langchain-community==0.3.19
langchain-core==0.3.41
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langsmith==0.1.147
llama-cloud==0.1.13
llama-cloud-services==0.6.3
llama-index==0.12.22
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.22
llama-index-embeddings-langchain==0.3.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.8
llama-index-legacy==0.9.48
llama-index-llms-openai==0.3.25
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.6
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-chroma==0.4.1
llama-index-vector-stores-qdrant==0.4.3
llama-parse==0.6.2
llamaindex-py-client==0.1.19
looseversion==1.3.0
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.7
mdurl==0.1.2
minijinja==1.0.16
mmh3==4.1.0
monotonic==1.6
motor==3.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
ollama==0.1.8
onnxruntime==1.18.0
open-clip-torch==2.24.0
openai==1.65.3
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.1
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
pillow==10.3.0
portalocker==2.8.2
posthog==3.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.6
prov==2.0.0
ptyprocess==0.7.0
pulsar-client==3.5.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydot==2.0.0
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.7.1
PyMuPDF==1.24.13
pymupdf4llm==0.0.17
pyparsing==3.1.2
pypdf==5.3.1
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-pptx==0.6.23
pytils==0.4.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
qdrant-client==1.8.2
ragas==0.1.7
rank-bm25==0.2.2
rdflib==7.0.0
redis==5.0.3
regex==2024.4.16
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.11.4
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
striprtf==0.0.26
sympy==1.12
tenacity==8.5.0
threadpoolctl==3.4.0
tiktoken==0.9.0
timm==0.9.16
tokenizers==0.21.0
tomli==2.0.1
tools==0.1.9
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
traits==6.3.2
transformers==4.49.0
triton==2.2.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
zstandard==0.23.0